Successfully troubleshooted accelerometer data ingress
Finished troubleshooting accelerometer. Updated register writes based
on what I gleaned from the datasheet. Apparently, the alternate address
0x53 does not work. This could be because of improper grounding,
because CS must be connected to ground to use 0x53, but I'm not sure.
Instead, I'm using the "main" device address 0x1D, which requires CS to
be connected to 3.3V (or HIGH). Due to fluctuations in the 3v3 supply,
it actually might be safer to directly connect CS to a digital out pin
and set the pin to HIGH. This acts as a more stable voltage source. I
will do that in the next revision.