Data storage breakthrough!
Instead of storing data as an int or double, store it in a char. Since
the
char can have a max length of 622 bytes, keep appending data until you
fill up the buffer, transmit the large set of data, erase the buffer
and then write to it again. It's my version of a circular buffer. To
that end, I've converted EKGRead to a char. I will experiment with
appending data to EKGRead tomorrow hopefully.