Human Testing 2
Commented out Particle.connect() for Human Testing 2. Wasn't able to
get the wireless router hooked up in ECB in time, so I decided to just
nix WiFi connectivity altogether for the test since we're not using it
anyway. I also added the voltage divider schematic for the temperature
sensor here.