SD Card status LED, other issues
Known issues - thermistorResistance and temperature are misreported. I
tried to implement cardErrorCode and cardErrorData but ran into some
issues. The source code declares cardErrorCode as a uint8_t, but uses
it as a bool elsewhere. I get errors when I try to use it as a bool. I
need to figure out how to translate the uint8_t to a bool that I can
use to control some status LEDs.