Fixed thermistor, time, data write
Manually setting time using setTime and inputting Unix Epoch Time. This
is the number of seconds since 1 January 1970. I'm doing this for our
special case where we don't have internet access during testing in ECB.
We may want to consider using this as well in actual device deployment,
since the device may be powered on in the absence of an internet
connection. Additionally, I made a minor modification to how data is
written to the card, to make it easier to plot later. Finally, I fixed
the inaccurate temperature data. I implemented the voltage divider
equation wrong - I had it set up such that the thermistor was wired to
ground, not the 10k resistor. Appears to be functioning fine now -
readings are far more reasonable. Yet to be verified.